+++ frontend/.env
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_PLACES_API_KEY = "AIzaSyB41KS49TJLvjP7W7xG5lvrSTC_EPMHQcQ"
