+++ frontend/Pages/MapScreen.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { View, Text } from 'react-native';
+
+export default function MapScreen({ navigation }: any) {
+    return (
+        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+            <Text
+                onPress={() => navigation.navigate('Home')}
+                style={{ fontSize: 26, fontWeight: 'bold' }}>Map Screen</Text>
+        </View>
+    );
+}
