+++ frontend/package.json
@@ -9,12 +9,19 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.10",
+    "@react-navigation/native": "^6.1.9",
+    "@types/react": "~18.2.14",
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-native": "0.72.6",
     "react-native-chart-kit": "^6.12.0",
-    "react-native-svg": "13.9.0"
+    "react-native-safe-area-context": "^4.7.3",
+    "react-native-svg": "^13.14.0",
+    "react-native-vector-icons": "^10.0.0",
+    "react-navigation": "^5.0.0",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
